Bug 27631: (follow-up) Some minor corrections
This patch makes some corrections to issues found during testing:

Some duplicate headings, a markup error in labeledMARCdetail.tt, a
change of the transfers heading from "Branch transfers" to "Item
transfers."

This patch also undoes the changes to the position of the toolbar on
several pages. I think those changes don't belong in this bug.

Signed-off-by: David Nind <david@davidnind.com>

Signed-off-by: Katrin Fischer <katrin.fischer.83@web.de>
Signed-off-by: Fridolin Somers <fridolin.somers@biblibre.com>