Bug 25387: The order of authority subfield is not sent over to bibliographic record
When creating an authority record there is the possibility to reorder
the subfield. However this order is not sent back to the bibliographic
records when using this authority in a bibliographic record.

This patch make the order of subfield be sent back to bibliographic
record when the autority is used.
Field that are fill are sent to the begining of the bibliographic record
in the same order as in the authority record.

Test plan:
1- Find or create an authority record and reorder some subfield
2- Use this authority in a bibliographic record
3- Observe that the order of the subfield is not transferred from the
   authority record to the bibliographic record
4- Apply the patch
5- Repeat step 1 and 2 again
6- Observe that the order of the subfield is now transferred from the
   authority record to the bibliographic record

https://bugs.koha-community.org/show_bug.cgi?id=22666

Signed-off-by: George Veranis <gveranis@dataly.gr>
Signed-off-by: Katrin Fischer <katrin.fischer.83@web.de>
Signed-off-by: Tomas Cohen Arazi <tomascohen@theke.io>